prevent init->up->stop when stop.txt exists
